Port CS:GO's own engine (cstrike15_src) instead of overlaying its game code
Replace the TF2-based engine tree with cstrike15_src (sources only; the
prebuilt Windows/Linux/OSX binaries are left out) and keep the iOS platform
pieces from the source-engine port:

- wscript: build every module straight from Valve's .vpc scripts via the
  new scripts/waifulib/vpc.py, following $Lib/$ImpLib links from the
  runtime modules. iOS builds as CS:GO's OSX64 configuration plus IOS=1,
  with Scaleform, Steam and CEG compiled out.
- protobuf 2.5.0 (the version CS:GO pins) built for host protoc and iOS;
  .proto files are regenerated every build.
- Crypto++ 5.6.1 built from external/.
- vphysics + ivp from the port (not part of the leak).
- nillerusr's thirdparty submodule moves to ios/thirdparty since CS:GO has
  its own thirdparty/; the unused lib submodule is dropped.
- CI caches ccache between runs and keeps going after compile errors.

+++ hammer/FileChangeWatcher.cpp
@@ -1,4 +1,4 @@
-//========= Copyright Valve Corporation, All rights reserved. ============//
+//====== Copyright © 1996-2005, Valve Corporation, All rights reserved. =======
 //
 // Purpose: 
 //
@@ -8,6 +8,7 @@
 #include "FileChangeWatcher.h"
 #include "tier1/utldict.h"
 #include "filesystem_tools.h"
+#include "vstdlib/vstrtools.h"
 
 
 CFileChangeWatcher::CFileChangeWatcher()
@@ -96,7 +97,7 @@ int CFileChangeWatcher::Update()
 				{
 					// Figure out what happened to this file.
 					WCHAR nullTerminated[2048];
-					int nBytesToCopy = min( (int)pNotify->FileNameLength, 2047 );
+					int nBytesToCopy = min( pNotify->FileNameLength, 2047 );
 					memcpy( nullTerminated, pNotify->FileName, nBytesToCopy );
 					nullTerminated[nBytesToCopy/2] = 0;
 					char ansiFilename[1024];
